Add review suggestions

+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/handlers/analytics/SynapseAnalyticsDataProvider.java
@@ -431,7 +431,7 @@ public class SynapseAnalyticsDataProvider implements AnalyticsDataProvider {
         return System.currentTimeMillis() - backendEndTime;
     }
 
-    public int getResponseSize() {
+    public long getResponseSize() {
         int responseSize = 0;
         if (buildResponseMessage == null) {
             Map<String,String> configs = APIManagerConfiguration.getAnalyticsProperties();
